[FIX] base: _name_search of state
When creating a new customer, the selection list for the state is empty
unless the country is filled in.

This is an issue since users usually fill the address fields in the
sequential order: street, city, state, zip and then country. Since no
state proposal is displayed, users tend to create a new state.

Introduced in 0b24dddb589c332a5723f7303a39108941ff8f6d.

opw-1941579

closes odoo/odoo#31273