[REF] website_slides: refactor the course reordering of slides and categories
Purpose
=======

This commit refactors the way slides and categories are reordered in the
frontend course view (training courses). The use of jqueryui's "sortable"
widget makes it easier and more user friendly to drag and drop both sections
and slides to reorganize your course.

Commit linked to task ID 1942852 and PR #31394.