[FIX] website_slides: fix call to set_completed
Currently fullscreen widget calls a user-specific route as public user if
user is not logged. Purpose of this commit is to avoid JS crash. Some
code cleaning is also done. This small feature will have to be cleaned more
in-depth in a near future.

Commit linked to task ID 1942852 and PR #31394.