[IMP] hr_contract: Improve display of contract related buttons
There is a button on the employee form that show the number of
contracts. This button is only accessible for hr manager,
but not for contract manager.

It's make more sense that this button is accessible for contract
manager. The contract manager can already see all contract from
the menu. Hr manager inherit rights of contract manager.

With this commit, the button is accessible with contract manager
rights and hr manager.

closes odoo/odoo#31415

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>