[FIX] sale: remove automatic invoice if invoice policy is set to delivered
Set the invoice policy to "ordered".
Then you can set to true the parameter "automatic_invoice", along with choosing
a mail template.
If you then set the invoice policy to "delivered",
"automatic_invoice" is rendered invisible.
However its value is not changed, so it is possible to have it set to true,
while the invoice policy is set to "delivered".

We change the parameter when settings are saved.

opw 1940641

closes odoo/odoo#31524